Refactor in preparation of custom search removal (issue #297).
