+++ app/lib/frontend/service_utils.dart
@@ -50,8 +50,7 @@ void initStorage(String projectId, authClient) {
 Future initSearchService() async {
   final searchService = await searchServiceViaApiKeyFromDb();
   registerSearchService(searchService);
-  registerScopeExitCallback(searchService.httpClient.close);
-  registerScopeExitCallback(searchService.searchServiceClient.close);
+  registerScopeExitCallback(searchService.close);
 }
 
 void initBackend({UIPackageCache cache}) {
