+++ app/test/frontend/handlers_test_utils.dart
@@ -193,18 +193,6 @@ class TemplateMock implements TemplateService {
 }
 
 class SearchServiceMock implements SearchService {
-  @override
-  // ignore: always_declare_return_types
-  get csearch => throw 'unexpected csearch';
-
-  @override
-  // ignore: always_declare_return_types
-  get httpClient => throw 'unexpected httpClient';
-
-  @override
-  // ignore: always_declare_return_types
-  get searchServiceClient => throw 'unexpected searchServiceClient';
-
   final Function searchFun;
 
   SearchServiceMock(this.searchFun);
@@ -213,4 +201,7 @@ class SearchServiceMock implements SearchService {
   Future<SearchResultPage> search(SearchQuery query, bool useService) async {
     return searchFun(query);
   }
+
+  @override
+  Future close() async => null;
 }
